move service-mirror into multicluster top level directory
